+++ tests/editor_api_compile.cpp
@@ -2,6 +2,7 @@
 #include <imkit/video.h>
 #include <imkit/cg.h>
 #include <imkit/preview.h>
+#include <imkit/preview_opengl3.h>
 #include <array>
 int main() {
     using namespace imkit;